+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## NEXT
+
+- Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_causal_graph` for
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not adding floating nodes
+  correctly.
+- Added `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.max_backward_lag` and
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.max_forward_lag` properties to
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` to return the absolute maximum backward and
+  forward time lag of the graph, respectively.
+- Fixed a bug with the property `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.maxlag` in
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` since it could give wrong information if the
+  future was included in the graph.
+
 ## 0.3.7
 
 - Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_adjacency_matrices` for